+++ app/src/views/index.ts
@@ -1,3 +1,4 @@
 export { CoordinatorDashboard } from "./CoordinatorDashboard";
 export { GuestApplicationTracker } from "./GuestApplicationTracker";
 export { HostApplicationTracker } from "./HostApplicationTracker";
+export { Home } from "./Home";
